host/requirements.txt
pyyaml>=6.0.1
pydantic>=2.7.0
pyserial>=3.5
requests>=2.32.0
plyer>=2.1.0
